+++ packages/button/src/Component.desktop.tsx
@@ -1,7 +1,7 @@
 import React, { forwardRef } from 'react';
 
 import { BaseButton } from './components/base-button';
-import { AnchorButtonProps, NativeButtonProps } from './typings';
+import { CommonButtonProps } from './typings';
 
 import defaultColors from './default.desktop.module.css';
 import styles from './desktop.module.css';
@@ -12,9 +12,7 @@ const colorStyles = {
     inverted: invertedColors,
 };
 
-export type ButtonDesktopProps = Partial<AnchorButtonProps | NativeButtonProps>;
-
-export const ButtonDesktop = forwardRef<HTMLAnchorElement | HTMLButtonElement, ButtonDesktopProps>(
+export const ButtonDesktop = forwardRef<HTMLAnchorElement | HTMLButtonElement, CommonButtonProps>(
     (restProps, ref) => (
         <BaseButton {...restProps} ref={ref} styles={styles} colorStylesMap={colorStyles} />
     ),