+++ packages/bank-card/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 5.4.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   masked-input@6.2.9
+
 ## 5.4.0
 
 ### Minor Changes