+++ packages/icon-button/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 6.3.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@9.0.6
+
 ## 6.3.0
 
 ### Minor Changes