+++ packages/toast/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 5.1.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   toast-plate@6.0.7
+
 ## 5.1.2
 
 ### Patch Changes